Remove fork support from branch name links
- GitHub changed some DOM in https://github.com/blog/2123-more-code-review-tools
The `.commit-id` is still matching elements on the page, but they are not anchors.
- I have tried changing the selector and making it future proof by checking the href attribute,
but actually they have changed links to the commits in a PR as well. They now use
/<source owner>/<source repo>/pull/<PR id>/commits/<commit sha>

That said I think we should remove the correct branch links across forks as it produces JS error.
We can restore it if we find another solution.

On top of that the links to a renamed fork are redirect unless there is another repo with the same name
so the branch links would work in most cases even without handling the renamed forks case.